Still trying to make things work
My previous approach didn't work on the prototype, `SER` still got
confused all the time. Now instead of having a fine-grained "lock" on
`SER`'s input/output mode, I'm trying a more coarse-grained approach
with `input_mode`.

Here, we begin each serial session with a single "empty" `INCLK` rising
edge that tells the MCU to stop all display refreshing activity. The MCU
then sets `SER` to input mode as soon as it can in the runloop.

On the other side of the serial wire, we have to wait a bit to make sure
we're good to go, and then go ahead with or 20 bits (for 4 digits) of
data.

After that, the `input_mode` flag goes back to `false` and normal
operations resume.

Also, use `INT0` interrupts instead of `PCINT0`. This allows us to save
ourselves a `pinisset()` call and have the interrupt called half as
much. `INT0` is on `PB2` so I had to re-wire my prototype a bit.

After much work, simulation is functional, will try soon on the
prototype. Crossing fingers...