Give INSER its own pin
The idea of sharing the `SER` pin with `INSER` was a bad idea. It might
be possible to make it work, but its over my current skill level.

`SR1`'s `RCLK` had its own pin, so I took it out and hard-wired it to
`SRCLK`. That makes the seg7 output code a bit uglier, but it frees a
pin.

This works on the prototype! finally!

There are some occasional glitches though, but they're much much less
frequent than ever before. I'll have to iron them out though...