Fix spurious DP display
Things worked well on simulation, but on the real thing, I would have my
DP lighting all the time, with the lighting level slightly differing
depending on the displayed glyph.

When designing the code, I thought that SER_DP briefly flickering when
sending data to the SR wouldn't significantly light DP because it was a
very small portion of the overall "SER_DP" position, which was, in the
vast majority of the time, kept HIGH.

It turns out that electricity is fast... This brief flickering was more
than enough to fully light DPs all the time.

This commit changes the logic to avoid this problem. The cost of this is
less overall "segment lighting time" because now, displays are fully
disabled for the whole time the SR is being updated. I initially wanted
to avoid this because I thought it would cause flickering, but it turns
out that it doesn't. So here we go.

I will also have to update icemu's assumptions about segments lighting
after brief pin changes.