Reverse digits (again)
It seems that I've turned my displays around without noticing in my
latest prototype.

But otherwise, I now have a prototype that work very well! It properly
receives input from FTDI, doesn't flicker and the board didn't need any
modification!

The only problem is that the board is a *tiny bit* too large on a corner
to fit in my enclosure. I could cut the corner, but there's an IC there.
There's enough place on the board to move it a bit, so I'll do that.

But first, I need to test the communication protocol when it originates
from another MCU (instead of a FTDI device).