Add checksum digit to communication protocol
My prototype was sometimes (but rarely), receiving the wrong number. I'm
not sure if it's due to bad soldering in my USB wire or voltage
irregularities or whatever else, but if I'm going to use those readings
for anything important, I have to be able to trust the the number
displayed is the good one.